Homework 4.4 done.

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -1,5 +1,7 @@
 package ru.hogwarts.school.service;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.model.Student;
@@ -11,21 +13,26 @@ import java.util.Set;
 @Service
 public class FacultyService {
     private final FacultyRepository repository;
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
 
     public FacultyService(FacultyRepository repository) {
         this.repository = repository;
     }
 
     public Faculty addFaculty(Faculty faculty) {
+        logger.info("Was invoked method for add faculty");
         return repository.save(faculty);
     }
 
     public Faculty getFacultyById(Long id) {
+        logger.info("Was invoked method for get faculty by id");
         return repository.findById(id).orElse(null);
     }
 
     public Faculty editFaculty(Faculty newFaculty) {
+        logger.info("Was invoked method for edit faculty");
         if (!repository.existsById(newFaculty.getId())) {
+            logger.error("Faculty not found by id=" + newFaculty.getId());
             return null;
         }
 
@@ -33,16 +40,25 @@ public class FacultyService {
     }
 
     public Faculty removeFaculty(long id) {
+        logger.info("Was invoked method for remove faculty");
+
         Faculty foundedFaculty = getFacultyById(id);
+        if (foundedFaculty == null) {
+            logger.error("Not found student by id=" + id);
+            return null;
+        }
         repository.delete(foundedFaculty);
+
         return foundedFaculty;
     }
 
     public List<Faculty> findByColor(String name, String color) {
+        logger.info("Was invoked method for find by color");
         return repository.findByColor(name, color);
     }
 
     public Set<Student> findStudentsByFaculty(Long facultyId) {
+        logger.info("Was invoked method for find students by faculty");
         return repository.findStudentsByFaculty(facultyId);
     }
 }
